server/requirements.txt
Flask
pytubefix
mutagen
moviepy==1.0.3
pyyaml
flask-cors